% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Radar coordinates with straight sides}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{Axis orientation.}

\item{start}{Start angle.}

\item{direction}{Direction sign.}
}
\description{
Radar coordinates with straight sides
}
\keyword{internal}
