% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_boundary_standardize}
\alias{.spi_boundary_standardize}
\title{Resolve ISO and country name columns from boundary source}
\usage{
.spi_boundary_standardize(geo_sf)
}
\arguments{
\item{geo_sf}{sf object from ArcGIS source.}
}
\description{
Resolve ISO and country name columns from boundary source
}
\keyword{internal}
