% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_fetch_boundaries}
\alias{.spi_fetch_boundaries}
\title{Fetch WB boundaries from ArcGIS API with local cache fallback}
\usage{
.spi_fetch_boundaries(resolution = "medium")
}
\arguments{
\item{resolution}{Character scalar: "medium" or "high".}
}
\description{
Fetch WB boundaries from ArcGIS API with local cache fallback
}
\keyword{internal}
