% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_geo_cache_dir}
\alias{.spi_geo_cache_dir}
\title{Geometry cache directory for SPI plotting maps}
\usage{
.spi_geo_cache_dir()
}
\description{
Geometry cache directory for SPI plotting maps
}
\keyword{internal}
