% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_geo_read_cache}
\alias{.spi_geo_read_cache}
\title{Read boundaries geometry cache}
\usage{
.spi_geo_read_cache(resolution)
}
\arguments{
\item{resolution}{Character map resolution key.}
}
\description{
Read boundaries geometry cache
}
\keyword{internal}
