% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_geo_write_cache}
\alias{.spi_geo_write_cache}
\title{Write boundaries geometry cache}
\usage{
.spi_geo_write_cache(geo_sf, resolution)
}
\arguments{
\item{geo_sf}{sf object.}

\item{resolution}{Character map resolution key.}
}
\description{
Write boundaries geometry cache
}
\keyword{internal}
