% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-inventory-cache.R
\name{.spi_inv_cache_dir}
\alias{.spi_inv_cache_dir}
\title{Resolve (and create if necessary) the spiR inventory cache directory}
\usage{
.spi_inv_cache_dir()
}
\value{
Character scalar: absolute path to the cache directory.
}
\description{
Resolve (and create if necessary) the spiR inventory cache directory
}
\keyword{internal}
