% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-map.R
\name{.spi_map_hover}
\alias{.spi_map_hover}
\title{Build map hover text}
\usage{
.spi_map_hover(dt, label, digits)
}
\arguments{
\item{dt}{data.table with map values.}

\item{label}{Label shown in tooltip.}

\item{digits}{Integer digits for value formatting.}
}
\description{
Build map hover text
}
\keyword{internal}
