% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-style.R
\name{.spi_theme_wb}
\alias{.spi_theme_wb}
\title{World Bank plot theme (replacement for wbplot::theme_wb)}
\usage{
.spi_theme_wb(chart_type = c("line", "map"))
}
\arguments{
\item{chart_type}{Either \code{"line"} (default) or \code{"map"}.}
}
\value{
A ggplot2 theme.
}
\description{
World Bank plot theme (replacement for wbplot::theme_wb)
}
\keyword{internal}
