% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spi_country}
\alias{spi_country}
\title{sqs_spi_country}
\usage{
spi_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Social Progress Index's Data.
If no argument is filed, all indicators will be displayed.
}
\examples{
mycountry <- spi_country()
mycountry <- spi_country(country = "Canada")
mycountry <- spi_country("Canada")

}
\seealso{
\code{\link{spi_indicator}} for the SPI's indicators and \code{\link{spi_data}} to collect data when you have both indicator and country code.
}
