% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-regions.R
\name{spi_plot_regions}
\alias{spi_plot_regions}
\title{Compare a SPI column across regions over time}
\usage{
spi_plot_regions(regions = NULL, value_col = "SPI.INDEX", version = "master")
}
\arguments{
\item{regions}{Optional character vector of region names.}

\item{value_col}{Character SPI column name.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Compare a SPI column across regions over time
}
