% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spiR_country-deprecated}
\alias{spiR_country-deprecated}
\alias{sqs_spi_country}
\title{sqs_spi_country (deprecated)}
\usage{
sqs_spi_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Social Progress Index's Data.
If no argument is filed, all indicators will be displayed. This function will be removed in the next version and replaced by the simpler function 'spir_country()'
}
\examples{
mycountry <- sqs_spi_country()
mycountry <- sqs_spi_country(country = "Canada")
mycountry <- sqs_spi_country("Canada")

}
\seealso{
\code{\link{spir_indicator}} for the SPI's indicators and \code{\link{spir_data}} to collect data when you have both indicator and country code.
}
