% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spir_country}
\alias{spir_country}
\title{spir_country}
\usage{
spir_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Social Progress Index's Data.
If no argument is filed, all indicators will be displayed.
}
\examples{
mycountry <- spir_country()
mycountry <- spir_country(country = "Canada")
mycountry <- spir_country("Canada")

}
\seealso{
\code{\link{spir_indicator}} for the SPI's indicators and \code{\link{spir_data}} to collect data when you have both indicator and country code.
}
