% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spiR_data-deprecated}
\alias{spiR_data-deprecated}
\alias{sqs_spi_data}
\title{sqs_spi_data (deprecated)}
\usage{
sqs_spi_data(
  country = spi_country,
  years = spi_year,
  indicators = spi_indicator
)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{years}{Years for which you want data.}

\item{indicators}{Indicators from the Social Progress Index.}
}
\value{
Data for the country, indicator and year requested.
}
\description{
This function allows you to find and display the Social Progress Index data according to the selected parameters.
If no arguments are filled, all data will be displayed. This function will be removed in the next version and replaced by the simpler function 'spir_data()'
}
\examples{
myData<- sqs_spi_data(country = c("USA", "FRA"), years = "2018", )
myData<- sqs_spi_data(country = c("USA", "FRA"), year = c("2018", "2019"), indicators = "SPI")
myData<- sqs_spi_data("USA", "2019", c("SPI", "FOW"))
myData<- sqs_spi_data(, "2018", )
myData<- sqs_spi_data("USA", "2017", )
myData<- sqs_spi_data("USA", , )
myData<- sqs_spi_data(, , )
myData<- sqs_spi_data()
}
\seealso{
\code{\link{spir_indicator}} for the SPI's indicator symbol and \code{\link{spir_country}} for the country's ISO code.
}
