% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spir_visual}
\alias{spir_visual}
\title{spir_visual}
\usage{
spir_visual(
  chart = "bar_1",
  indicator = "SPI",
  years = max(SPI_data$year),
  title = TRUE
)
}
\arguments{
\item{chart}{Type of charts.}

\item{indicator}{An indicator from the Social Progress Index.}

\item{years}{A chosen year}

\item{title}{Chart title, set by default to TRUE}
}
\value{
Chosen Graph
}
\description{
This function allows you to create 2 types of visuals: bar and line charts.
}
\examples{
spir_visual(chart = "bar_1", indicator = "SPI", years = "2020")
}
