% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiCountry.R
\name{sqs_spi_country}
\alias{sqs_spi_country}
\title{sqs_spi_country}
\usage{
sqs_spi_country(COUNTRY)
}
\arguments{
\item{COUNTRY}{The name of the country starting with a capital letter.}
}
\value{
A country code
}
\description{
This function allows you to find and search the right country code associated with the Social Progress Index's Data.

If no argument is filed, all countries will be displayed.
}
\examples{
sqs_spi_country("Canada")
AllCountries <- sqs_spi_country()
}
\seealso{
\code{\link{sqs_spi_symbol}} for the SPI's indicator symbol and \code{\link{sqs_spi_data}} to collect data when you have both the indicator and the country code.
}
