% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiData.R
\name{sqs_spi_data}
\alias{sqs_spi_data}
\title{sqs_spi_data}
\usage{
sqs_spi_data(COUNTRY_CODE, INDICATOR_CODE, YEAR)
}
\arguments{
\item{COUNTRY_CODE}{A country code associated to the Social Progress Index.}

\item{INDICATOR_CODE}{A Indicator code associated}

\item{YEAR}{A year for which you want data}
}
\value{
Data for the country, indicator and year requested
}
\description{
This function allows you to find and display the Social Progress Index data according to the selected parameters.

If no arguments are filed, all data will be displayed.
}
\examples{
myData <- sqs_spi_data("CAN","AIC_4","2018")
DataFromAllYears <- sqs_spi_data("CAN","AIC_4",)
DataFromAllCountries <- sqs_spi_data(,"AIC_4","2018")
DataFromAllIndicators <- sqs_spi_data("CAB",,"2018")
}
\seealso{
\code{\link{sqs_spi_symbol}} for the SPI's indicator symbol and \code{\link{sqs_spi_country}} for the SPI's country code.
}
