% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{sqs_spi_indicator}
\alias{sqs_spi_indicator}
\title{sqs_spi_symbol}
\usage{
sqs_spi_indicator(indicators)
}
\arguments{
\item{indicators}{An indicator from the Social Progress Index.}
}
\value{
Indicator code from the Social Progress Index.
}
\description{
This function allows you to find and search the right indicator code from the Social Progress Index you want to use.
If no argument is filed, all indicators will be displayed.

This function allows you to find and search the right indicator code from the Social Progress Index you want to use.
}
\examples{
mysymbol <- sqs_spi_symbol()
mysymbol <- sqs_spi_symbol(indicators = "mortality")
mysymbol <- sqs_spi_symbol("mortality")

}
