% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiSymbol.R
\name{sqs_spi_symbol}
\alias{sqs_spi_symbol}
\title{sqs_spi_symbol}
\usage{
sqs_spi_symbol(INDICATOR)
}
\arguments{
\item{INDICATOR}{A social indicator from the Social Progress Index.}
}
\value{
Indicator code from the Social Progress Index.
}
\description{
This function allows you to find and search the right indicator code from the Social Progress Index you want to use.

If no argument is filed, all indicators will be displayed.
}
\examples{
sqs_spi_symbol("electricity")
AllIndicators <- sqs_spi_symbol()
}
\seealso{
\code{\link{sqs_spi_country}} for the SPI's country code and \code{\link{sqs_spi_data}} to collect data when you have both the indicator and the country code.
}
