% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-style.R
\name{.spi_scale_color_wb_d}
\alias{.spi_scale_color_wb_d}
\title{Discrete WB colour scale (replacement for wbplot::scale_color_wb_d)}
\usage{
.spi_scale_color_wb_d(n = length(SPI_WB_CAT), ...)
}
\arguments{
\item{n}{Integer. Number of discrete levels to support.}

\item{...}{Passed to \code{\link[ggplot2:scale_colour_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete WB colour scale (replacement for wbplot::scale_color_wb_d)
}
\keyword{internal}
