% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-style.R
\name{.spi_scale_fill_wb_d}
\alias{.spi_scale_fill_wb_d}
\title{Discrete WB fill scale (replacement for wbplot::scale_fill_wb_d)}
\usage{
.spi_scale_fill_wb_d(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete WB fill scale (replacement for wbplot::scale_fill_wb_d)
}
\keyword{internal}
