% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-wrappers.R
\name{metadata_pillars}
\alias{metadata_pillars}
\title{Retrieve SPI pillar metadata}
\usage{
metadata_pillars(version = "master")
}
\arguments{
\item{version}{Character. Branch name in SPI repository. Defaults to
\code{"master"}.}
}
\value{
A \code{data.table} with pillar-level metadata.
}
\description{
Convenience wrapper that returns the pillar block from \code{\link[=metadata]{metadata()}}.
}
\examples{
\dontrun{
metadata_pillars()
}
}
\seealso{
\code{\link[=metadata]{metadata()}}, \code{\link[=metadata_dimensions]{metadata_dimensions()}}
}
