% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\docType{data}
\name{spi_data}
\alias{spi_data}
\title{sqs_spi_data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1434 rows and 72 columns.}
\usage{
spi_data
}
\arguments{
\item{country}{Countries' ISO code.}

\item{years}{Years for which you want data.}

\item{indicators}{Indicators from the Social Progress Index.}
}
\value{
Data for the country, indicator and year requested
}
\description{
This function allows you to find and display the Social Progress Index data according to the selected parameters.
If no arguments are filed, all data will be displayed.
}
\examples{
mydata2 <- sqs_spi_data(, , )
}
\keyword{datasets}
