% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{spi_data}
\alias{spi_data}
\title{spi_data}
\usage{
spi_data(
  country = data_long_country,
  years = data_long_year,
  indicators = data_long_indicator
)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{years}{Years for which you want data.}

\item{indicators}{Indicators from the Social Progress Index.}
}
\value{
Data for the country, indicator and year requested
}
\description{
This function allows you to find and display the Social Progress Index data according to the selected parameters.
If no arguments are filled, all data will be displayed.
}
\examples{
myData<- spi_data(country = c("USA", "FRA"), years = "2018", )
myData<- spi_data(country = c("USA", "FRA"), year = c("2018", "2019"), indicators = "SPI")
myData<- spi_data("USA", "2019", c("SPI", "FOW"))
myData<- spi_data(, "2018", )
myData<- spi_data("USA", "2017", )
myData<- spi_data("USA", , )
myData<- spi_data(, , )
myData<- spi_data()


}
\seealso{
\code{\link{spi_indicator}} for the SPI's indicator symbol and \code{\link{spi_country}} for the country's ISO code.
}
