% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-country-vs-region.R
\name{spi_plot_country_vs_region}
\alias{spi_plot_country_vs_region}
\title{Plot a country against its regional average over time}
\usage{
spi_plot_country_vs_region(
  country,
  value_col = "SPI.INDEX.PIL1",
  version = "master"
)
}
\arguments{
\item{country}{Character scalar with country name or ISO3 code.}

\item{value_col}{Character SPI column name.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Plot a country against its regional average over time
}
