% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-pillars.R
\name{spi_plot_pillars}
\alias{spi_plot_pillars}
\title{Plot SPI pillar trajectories over time for one country}
\usage{
spi_plot_pillars(
  country,
  pillars = paste0("SPI.INDEX.PIL", 1:5),
  version = "master"
)
}
\arguments{
\item{country}{Character scalar with country name or ISO3 code.}

\item{pillars}{Character vector with pillar columns.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Plot SPI pillar trajectories over time for one country
}
