% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-radar.R
\name{spi_plot_radar}
\alias{spi_plot_radar}
\title{Radar chart of SPI pillars for one country vs region average}
\usage{
spi_plot_radar(country, year, version = "master")
}
\arguments{
\item{country}{Character scalar with country name or ISO3 code.}

\item{year}{Integer year to display.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Radar chart of SPI pillars for one country vs region average
}
