% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-region-pillars.R
\name{spi_plot_region_pillars}
\alias{spi_plot_region_pillars}
\title{Plot region pillar trajectories over time}
\usage{
spi_plot_region_pillars(
  region,
  weighted = TRUE,
  pillars = paste0("SPI.INDEX.PIL", 1:5),
  version = "master"
)
}
\arguments{
\item{region}{Character scalar with WB region name.}

\item{weighted}{Logical. TRUE for population-weighted average.}

\item{pillars}{Character vector of pillar columns.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Plot region pillar trajectories over time
}
