% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-plot-trend.R
\name{spi_plot_trend}
\alias{spi_plot_trend}
\title{Compare any SPI column over time across countries}
\usage{
spi_plot_trend(countries, value_col = "SPI.INDEX", version = "master")
}
\arguments{
\item{countries}{Character vector of country names or ISO3 codes.}

\item{value_col}{Character SPI column name.}

\item{version}{Character SPI branch.}
}
\value{
A ggplot object.
}
\description{
Compare any SPI column over time across countries
}
