% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-github.R
\name{spi_versions}
\alias{spi_versions}
\title{List available SPI versions (branches)}
\usage{
spi_versions()
}
\value{
A sorted character vector of branch names. \code{"master"} (the
latest stable version) is always included.
}
\description{
Queries the GitHub API for all branches in the World Bank SPI repository.
Each branch corresponds to a version of the SPI data that can be passed
as the \code{version} argument to \code{\link[=spi_get]{spi_get()}}.
}
\details{
If the repository has more than 100 branches, the list may be incomplete
due to GitHub API pagination limits; a warning is issued in this case.
}
\examples{
\dontrun{
spi_versions()
}

}
\seealso{
\code{\link[=spi_get]{spi_get()}}
}
