% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR-package.R
\docType{package}
\name{spiR-package}
\alias{spiR}
\alias{spiR-package}
\title{spiR: Statistical Performance Indicators Data Access}
\description{
Access World Bank Statistical Performance Indicators (SPI) data from
GitHub. Provides \code{\link[=spi_get]{spi_get()}}, \code{\link[=spi_data]{spi_data()}}, \code{\link[=spi_index]{spi_index()}}, and
\code{\link[=spi_aggregates]{spi_aggregates()}} to retrieve and filter the core SPI output datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WB-DECIS/spiR}
  \item Report bugs at \url{https://github.com/WB-DECIS/spiR/issues}
}

}
\author{
\strong{Maintainer}: World Bank DECDG \email{decdg@worldbank.org}

}
\keyword{internal}
