% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{sqs_spi_country}
\alias{sqs_spi_country}
\title{sqs_spi_country}
\usage{
sqs_spi_country(country)
}
\arguments{
\item{country}{The name of the country starting with a capital letter.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Social Progress Index's Data.
If no argument is filed, all indicators will be displayed.
}
