% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiR.R
\name{sqs_spi_data}
\alias{sqs_spi_data}
\title{sqs_spi_data}
\usage{
sqs_spi_data(country = data_long_country, years = data_long_year,
  indicators = data_long_indicator)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{years}{Years for which you want data.}

\item{indicators}{Indicators from the Social Progress Index.}
}
\value{
Data for the country, indicator and year requested
}
\description{
This function allows you to find and display the Social Progress Index data according to the selected parameters.
If no arguments are filed, all data will be displayed.
}
